+++ public/class-tmsm-frontend-optimizations-public.php
@@ -516,7 +516,7 @@ class Tmsm_Frontend_Optimizations_Public {
 	}
 
 	/**
-	 * Gravity Forms: Filters the field content.
+	 * Gravity Forms: Add autocomplete off to number field type.
 	 *
 	 * @param string $content    The field content.
 	 * @param array  $field      The Field Object.
@@ -528,6 +528,24 @@ class Tmsm_Frontend_Optimizations_Public {
 
 		if ($field['type'] == 'number') {
 			$content = str_replace('<input', '<input autocomplete="off"', $content);
+		}
+
+		return $content;
+	}
+
+	/**
+	 * Gravity Forms: Add phone format to phone field type.
+	 *
+	 * @param string $content    The field content.
+	 * @param array  $field      The Field Object.
+	 * @param string $value      The field value.
+	 * @param int    $lead_id The entry ID.
+	 * @param int    $form_id The form ID.
+	 */
+	public function gravityforms_field_content_phoneformat($content, $field, $value, $lead_id, $form_id){
+
+		if ($field['type'] == 'phone') {
+			$content = str_replace('<input', '<input data-phoneformat="'.($field['phoneFormat'] ?? '').'"', $content);
 		}
 
 		return $content;